code/python/apps/composite_indexes/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
arrow==1.0.3
    # via -r requirements.in
azure-core==1.13.0
    # via azure-cosmos
azure-cosmos==4.2.0
    # via -r requirements.in
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
docopt==0.6.2
    # via -r requirements.in
idna==2.10
    # via requests
python-dateutil==2.8.1
    # via arrow
requests==2.25.1
    # via
    #   -r requirements.in
    #   azure-core
six==1.15.0
    # via
    #   azure-core
    #   azure-cosmos
    #   python-dateutil
urllib3==1.25.11
    # via requests
